+++ tvos/BrightcovePlayerSDK.framework/Headers/BCOVCatalogService.h
@@ -42,9 +42,12 @@
 @end
 
 /**
- * The BCCatalog class provides asynchronous methods for retrieving information
+ * Deprecated: Use BCOVPlaybackService instead.
+ *
+ * The BCOVCatalogService class provides asynchronous methods for retrieving information
  * about videos and playlists from Brightcove’s Media API services.
  */
+DEPRECATED_MSG_ATTRIBUTE("Use BCOVPlaybackService instead")
 @interface BCOVCatalogService : NSObject
 
 /**